+++ packages/plugin/src/index.ts
@@ -62,6 +62,7 @@ import preferShortcutReturn from './rules/prefer-shortcut-return';
 import preferStringRaw from './rules/prefer-string-raw';
 import requireBddSpec from './rules/require-bdd-spec';
 import requireInterfacePrefix from './rules/require-interface-prefix';
+import requireJsdocAnonymousFunctions from './rules/require-jsdoc-anonymous-functions';
 import requireJsdocFunctions from './rules/require-jsdoc-functions';
 import requireOptionalChaining from './rules/require-optional-chaining';
 import requireReadonlyProps from './rules/require-readonly-props';
@@ -126,6 +127,7 @@ const rules: Record<string, unknown> = {
   'prefer-string-raw': preferStringRaw,
   'require-interface-prefix': requireInterfacePrefix,
   'require-bdd-spec': requireBddSpec,
+  'require-jsdoc-anonymous-functions': requireJsdocAnonymousFunctions,
   'require-jsdoc-functions': requireJsdocFunctions,
   'require-optional-chaining': requireOptionalChaining,
   'require-readonly-props': requireReadonlyProps,